requirements.txt
# Core Dependencies
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.0.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
Pillow>=8.0.0

# NLP & Evaluation
nltk>=3.6.0
rouge-score>=0.1.2
Levenshtein>=0.20.0

# API & Tools
openai>=0.27.0
tqdm>=4.62.0

# Vision & Document Processing
vidore-benchmark>=0.1.0
pymupdf>=1.23.0  # For PDF processing

# Optional but recommended
matplotlib>=3.3.0
seaborn>=0.11.0
jupyter>=1.0.0
ipywidgets>=7.6.0
